extract cite into package
